Extract app loader
