+++ packages/app/src/cli/models/app/app.test-data.ts
@@ -1,5 +1,5 @@
 import {App} from './app.js'
-import {UIExtension} from './extensions'
+import {UIExtension} from './extensions.js'
 
 export function testApp(app: Partial<App> = {}): App {
   return {
